2-Run_Level_Configuration/Maths_kids_agent/pyproject.toml
[project]
name = "maths-kids-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
]
